[chore][exporter/datadog] skip known data race
